Split out connector trait and fns into own module
